separate gui and experiment folders
