+++ customxepr/config/base.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 """
 Base configuration management
