+++ customxepr/gui/manager_ui_linux.ui
