+++ customxepr/gui/manager_ui_macos.ui
