+++ customxepr/gui/resources/aborted@2x.icns
