+++ customxepr/gui/resources/logo@2x.png
