+++ customxepr/gui/resources/running@2x.icns
