+++ customxepr/gui/scientific_spinbox.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
 This file contains a wrapper to display the SpinBox in scientific way
